Create the code
This is a assignment based on the origin code from lecture and adding a start menu with "new game", "load game" and "quit" choices as well as some functions, such as "save game", "get a wanted inventory" or "go any wanted position". Feel free to modify and build yourself a classic python game!